Add draft version of log, exp, and geodesic solvers
